just sync

+++ contracts/Events.sol
@@ -5,5 +5,5 @@ pragma solidity ^0.8.0;
 interface Events {
     event NewToken(address submitter, address tokenAddr, uint16 tokenId);
     event RegisterUser(address ethAddr, uint16 userId, bytes32 bjjPubkey);
-    event Deposit(uint16 tokenId, bytes32 to, uint256 amount);
+    event Deposit(uint16 tokenId, bytes32 to, uint128 amount);
 }
